Changed updating users profile to fetch current user using current authentication to check if he is trying to edit his own profile

+++ Spring/src/main/java/pl/ttsw/GameRev/service/WebsiteUserService.java
@@ -54,8 +54,14 @@ public class WebsiteUserService {
     }
 
     public WebsiteUser updateUserProfile(String username, UpdateWebsiteUserDTO request) throws BadRequestException {
-        WebsiteUser user = websiteUserRepository.findByUsername(username);
+        WebsiteUser user = getCurrentUser();
 
+        if(user == null){
+            throw new BadRequestException("You need to login first");
+        }
+        if(!username.equals(user.getUsername())){
+            throw new BadRequestException("You can only edit your own profile");
+        }
         if (!passwordEncoder.matches(request.getCurrentPassword(), user.getPassword())) {
             throw new BadRequestException("Passwords do not match");
         }
@@ -89,6 +95,12 @@ public class WebsiteUserService {
         MultipartFile file = profilePictureDTO.getProfilePicture();
         WebsiteUser user = websiteUserRepository.findByUsername(username);
 
+        if(user == null){
+            throw new BadRequestException("This user does not exist");
+        }
+        if(user != getCurrentUser()){
+            throw new BadRequestException("You can only edit your own profile picture");
+        }
         if (user.getProfilepic() != null && !user.getProfilepic().isEmpty()) {
             Path oldFilepath = Paths.get(user.getProfilepic());
             Files.deleteIfExists(oldFilepath);
